Moved the sensor package to the util.sensors package and refactored other classes with it.
