+++ core/src/cg/group4/util/sensors/AccelerationStatus.java
@@ -1,4 +1,6 @@
-package cg.group4.sensor;
+package cg.group4.util.sensors;
+
+import cg.group4.util.sensors.AccelerationState;
 
 /**
  * Interface which will be the core module wrapper for the Accel Library Android.
